backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.10.1
pydantic-settings==2.7.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.3.3
python-multipart==0.0.20
alembic==1.14.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.27.2
